+++ magento.sh
@@ -13,6 +13,7 @@ URL_MAILCATCHER="http://localhost:1096/"
 header="bin/tests/"
 pathPreFile=${header}000*/0_init/*.js
 pathDir=${header}0*
+containerMG2=hipay-fullservice-sdk-magento2_web_1
 
 if [ "$1" = '' ] || [ "$1" = '--help' ]; then
     printf "\n                                                      "
@@ -34,32 +35,36 @@ if [ "$1" = 'init' ]; then
         rm -Rf log/ web/
         docker-compose -f docker-compose.dev.yml build
         COMPOSE_HTTP_TIMEOUT=200 docker-compose -f docker-compose.dev.yml up -d
-        docker cp hipayfullservicesdkmagento2_web_1:/var/www/html/magento2 web/
+        docker cp $containerMG2:/var/www/html/magento2 web/
         docker-compose -f docker-compose.dev.yml logs -f
     else
         echo "Put your credentials in auth.env and hipay.env before start update the docker-compose.dev to link this files"
     fi
+elif [ "$1" = 'kill' ]; then
+    docker-compose -f docker-compose.dev.yml stop
+    docker-compose -f docker-compose.dev.yml rm -fv
+    rm -Rf log/ web/
 elif [ "$1" = 'start_https' ]; then
     docker-compose -f docker-compose.dev-https.yml up -d --build
 elif [ "$1" = 'restart' ]; then
     docker-compose -f docker-compose.dev.yml stop
     docker-compose -f docker-compose.dev.yml up -d
 elif [ "$1" = 'static' ]; then
-    docker exec hipayfullservicesdkmagento2_web_1 rm -Rf /var/www/html/magento2/pub/static/frontend/Magento/luma/en_US/HiPay_FullserviceMagento/
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 php bin/magento setup:static-content:deploy -t Magento/luma
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 php bin/magento c:c
+    docker exec $containerMG2 rm -Rf /var/www/html/magento2/pub/static/frontend/Magento/luma/en_US/HiPay_FullserviceMagento/
+    docker exec $containerMG2 gosu magento2 php bin/magento setup:static-content:deploy -t Magento/luma
+    docker exec $containerMG2 gosu magento2 php bin/magento c:c
 elif [ "$1" = 'di' ]; then
-    docker exec hipayfullservicesdkmagento2_web_1 rm -Rf /var/www/html/magento2/var/cache /var/www/html/magento2/var/di /var/www/html/magento2/var/generation /var/www/html/magento2/var/page_cache
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 php bin/magento setup:di:compile
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 php bin/magento c:c
+    docker exec $containerMG2 rm -Rf /var/www/html/magento2/var/cache /var/www/html/magento2/var/di /var/www/html/magento2/var/generation /var/www/html/magento2/var/page_cache
+    docker exec $containerMG2 gosu magento2 php bin/magento setup:di:compile
+    docker exec $containerMG2 gosu magento2 php bin/magento c:c
 elif [ "$1" = 'command' ]; then
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 php bin/magento $2
+    docker exec $containerMG2 gosu magento2 php bin/magento $2
 elif [ "$1" = 'l' ]; then
     docker-compose -f docker-compose.dev.yml logs -f
 elif [ "$1" = 'install' ]; then
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 bin/magento module:enable --clear-static-content HiPay_FullServiceMagento
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 bin/magento setup:upgrade
-    docker exec hipayfullservicesdkmagento2_web_1 gosu magento2 bin/magento c:c
+    docker exec $containerMG2 gosu magento2 bin/magento module:enable --clear-static-content HiPay_FullServiceMagento
+    docker exec $containerMG2 gosu magento2 bin/magento setup:upgrade
+    docker exec $containerMG2 gosu magento2 bin/magento c:c
 elif [ "$1" = 'test' ]; then
 
     rm -rf bin/tests/errors/*